Update to work with tags and Smithy CLI
This commit updates the plugin to add support for projection and JAR
tags when building projections. This also updates the tasks to have a
common hierarchy where possible, but utilize the Smithy CLI abstraction
for building sources when building the JAR (allowing for that logic to
be a single place rather than reimplemented everywhere that Smithy build
tooling is introduced).