Run CLI in thread by default
Rather than always run in a process, this change allows the Smithy CLI
to be run in a thread using a custom class loader. This will probably
work in every case, but just in case, I added an option to force the CLI
to run in a separate process using the "fork" option.

On my machine, this change had the effect of shaving on ~1 second of time to
run a normal build. You also get much better stack traces when something
goes wrong since it's not run in a different process.