Fix build graph issue
Smithy needs to run after compileJava so that dependencies in the same
project are built first. It also needs to run before processResources so
that artifacts created by Smithy are considered part of the generated
JAR.

This has been tested using the integration tests in this package as well
as using the awslabs/smithy repo with various states (clean build and
cleaning individual projects).