Scan for smithy-cli plugin when building
This commit updates the Gradle plugin to scan buildScript dependencies
as a last resort for determining which version of the smithy-cli to use
when building the model. Since the gradle plugin has a dependency on
smithy-cli, this should always return the version of the smithy-cli that
the gradle plugin was built against (not that other version can be used
for example, like when there's an explicit dependency on smithy-cli).

This change should help with a getting started experience, and the
warning message should guide developers to add an explicit dependency on
smithy-model if they are producing JARs.