Update Gradle plugin to work with new Smithy CLI
The Smithy CLI was updated to change some of its output, and remove
harder to test behaviors (like using a global variable to determine
where to send output). This commit updates the Gradle plugin to work
with these updates, but the plugin will also continue to work with older
versions of the CLI too.