Make failed tests something we can actually fix
When tests failed in GitHub actions, we had no output to act on. This
change ensures that stderr/stdout are printed by each integration test,
and then uses a logging plugin to dump out this captured output on
failure so we can, ya know, understand how to fix weird platform
specific issues if they come up in integration tests.