Support implicit dependencies
This plugin has two tasks that produce output, `SmithyBuild` and
`SmithyBuildJar`. Both of these tasks model their outputs as a
`java.io.File`, which results in a few problems:

1. The output locations are eagerly evaluated, which can introduce race
   conditions if output locations are overridden through the task itself
   or the `SmithyExtension`
2. Gradle cannot track which tasks are responsible for producing which
   collections of Smithy files, which means that explicit dependencies
   may need to be introduced by calling `dependsOn`

This commit leaves the current APIs in place for compatibility and adds
new ones that support lazy configuration and implicit task dependency
tracking. For more information on these Gradle features, see:

* https://docs.gradle.org/current/userguide/lazy_configuration.html
* https://docs.gradle.org/7.4.2/userguide/validation_problems.html#implicit_dependency