Pass logging, debug, and stacktrace params to CLI
The Smithy CLI supports --debug, --logging, and --stacktrace
parameters. This commit updates the Gradle pluging to pass
these arguments through to the CLI if they were provided to
Gradle. All versions of the Smithy CLI support these arguments
after 1.0, so this is a safe change regardless of the resolved
CLI version.