Update task ordering (#103)
Updates task ordering. The Format task now executes before the build task to ensure any files written to an output jar will be formatted. Also, the processResources now runs before the compile task to ensures that smithy output files are in the correct output resource location before the compile task is executed so they are available to the compile task for any annotation processors.